Refactor, first version laser tower

+++ Assets/Scripts/Items/Bullets/AoEBulletProfile.cs
@@ -2,11 +2,11 @@
 {
     public float Speed { get; }
 
-    public int Damage { get; }
+    public float Damage { get; }
 
     public float Range { get; }
 
-    public AoEBulletProfile(float speed, int damage, float range)
+    public AoEBulletProfile(float speed, float damage, float range)
     {
         Speed = speed;
         Damage = damage;
